1) There has been the long-standing issue of making the high-level methods for the access of TBB time-series dat in HDF5 format part of the DAL; originally - when I started with them - they were part of the CR-Tools. However with the ongoing upgrades to the classes this has been the perfect opportunity to finally collect all the functionality within the DAL: the classes LOFAR_DipoleDataset, LOFAR_StationGroup and LOFAR_Timeseries now have been moved to DAL::TBB_DipoleDataset, DAL::TBB_StationGroup and DAL::TBB_Timeseries - a number of functional upgrades have been done already do allow recursive access to the attributes. Resulting adjustments in the CR-Tools source code have been made already.
2) Inspecting the build logs John sent in, I think I have been able to resolve the problem of num_util not
building due to missing the NumPy header files - this is an issue which potentially should show up across
all systems unless there is a system-wide NumPy installation in a place where the configuration scripts can
pick it up. Since include paths are set during configuration time, the header files must be search for also
within the source code tree, because they will be showing up in the installation location only afterwards
(essentially it is that one cannot prove for something which has not been created yet).


git-svn-id: http://usg.lofar.org/svn/code/trunk@2045 e3a606d5-b720-0410-b6ad-cb58f335abf8