Bugfixes.
We're providing a default GVDS file for each recipe that needs one, with the
filename given in the config file. This is handy for testing, but is it really
sensible long term? Bear in mind that things like the imager will be run
multiple times with different GVDS files in one job.

git-svn-id: http://usg.lofar.org/svn/code/trunk@3468 e3a606d5-b720-0410-b6ad-cb58f335abf8