Add a search path for python recipes.
This is slightly hacked in at the moment, since it's not clear how exactly it
should interface with the outside world.

For now, path defaults to '.' (ie, cwd) for a straightforward cuisine recipe,
but should be adjusted in pipeline.cfg.

Note each LOFARrecipe reads the config file and sets this freshly for itself
-- should it be a more straightforward member of the input array/option
parser?

git-svn-id: http://usg.lofar.org/svn/code/trunk@3632 e3a606d5-b720-0410-b6ad-cb58f335abf8