Cleaned up hftools build system.
Changed build system for hftools to be selfcontained and work for multiple input files.
Also separated Python modules from scripts and added install targets for both.
These targets create symbolic links so changing the source code is immediately reflected in the imported modules, as is expected for Python code.

In contrast to modules, scripts should always start with "#! /usr/bin/env python" and have executable permissions.
Furthermore all nessesary modules for running a script should always be imported within the script itself so the script is selfcontained and can be run with ./scriptname.py.
This was changed for some, but not all, scripts.

git-svn-id: http://usg.lofar.org/svn/code/trunk@4980 e3a606d5-b720-0410-b6ad-cb58f335abf8