correct the coordinates system -> now all in the LOPES coord. system!
The dictionary file gives only info on the id (azimuth and zenith are
dummy values!).
For the simulations, the distance in the shower coordinates system as to
be calculated considering Az from KASCADE (the same value used for the
REAS simulations!!). Both KASCADE and LOPES Az is null in the North!


git-svn-id: http://usg.lofar.org/svn/code/trunk@5249 e3a606d5-b720-0410-b6ad-cb58f335abf8