Update to LPPS search script (refactor, candidate P,F histograms, bugfixes).
Bugfix to declination coordinate handling was important, the rest of the bug
fixes were to harmless bugs. The candidates are now put in a histogram by
perdiod and by frequency before and after sifting the candidates. In general
I have been refactoring code to move as much as possible out of the main
search script.

git-svn-id: http://usg.lofar.org/svn/code/trunk@7156 e3a606d5-b720-0410-b6ad-cb58f335abf8