Correctly link see against libX11.
Previously, the following command line was generated to build the see
exectuable:

  /usr/bin/c++     -L/usr/lib/libX11.so -L/usr/lib -lSM -lICE -lXt [etc]

This has the incorrect syntax for linking against libX11 (should be -lX11).
However, libXt is correctly linked, and, in turn, it links against libX11.

On older platforms (including that running on lfe001) that was fine: the
incorrect syntax was ignored, and libX11 was linked indirectly through libXt.

On newer platforms, indirect linking is disabled. The build will therefore
fail, since symbols in libX11 cannot be resolved:

  /usr/bin/ld: CMakeFiles/see.dir/regard.cpp.o: undefined reference to symbol 'XGetWindowAttributes'
  /usr/bin/ld: note: 'XGetWindowAttributes' is defined in DSO //usr/lib64/libX11.so.6 so try adding it to the linker command line

This change was made in Debian GCC 4.4.5-11, dated 15 Feb 2011. It will
(presumably) propagate downstream to Ubuntu in an upcoming release. It is also
planned for Fedora.

This patch fixes the issue by corectly linking against libX11, thereby
avoiding future problems. Phew.

See the following for details:

http://wiki.debian.org/ToolChain/DSOLinking
https://fedoraproject.org/wiki/UnderstandingDSOLinkChange

git-svn-id: http://usg.lofar.org/svn/code/trunk@7470 e3a606d5-b720-0410-b6ad-cb58f335abf8